replace parquet.{Type}Reader and parquet.{Type}Writer with parquet.RequiredReader[T] and parquet.RequiredWriter[T]
