+++ column_buffer_default.go
@@ -646,10 +646,12 @@ var (
 	_ io.Writer = (*floatColumnBuffer)(nil)
 	_ io.Writer = (*doubleColumnBuffer)(nil)
 
-	_ BooleanWriter = (*booleanColumnBuffer)(nil)
-	_ Int32Writer   = (*int32ColumnBuffer)(nil)
-	_ Int64Writer   = (*int64ColumnBuffer)(nil)
-	_ Int96Writer   = (*int96ColumnBuffer)(nil)
-	_ FloatWriter   = (*floatColumnBuffer)(nil)
-	_ DoubleWriter  = (*doubleColumnBuffer)(nil)
+	_ BooleanWriter           = (*booleanColumnBuffer)(nil)
+	_ Int32Writer             = (*int32ColumnBuffer)(nil)
+	_ Int64Writer             = (*int64ColumnBuffer)(nil)
+	_ Int96Writer             = (*int96ColumnBuffer)(nil)
+	_ FloatWriter             = (*floatColumnBuffer)(nil)
+	_ DoubleWriter            = (*doubleColumnBuffer)(nil)
+	_ ByteArrayWriter         = (*byteArrayColumnBuffer)(nil)
+	_ FixedLenByteArrayWriter = (*fixedLenByteArrayColumnBuffer)(nil)
 )
